Use with_color scope to filter by color
Rails' enum querying has come back to bite us again :-(
If you pass a string for an enum query than it will coerce it to an
integer which will usually be 0. Since enums 0 based indexing, this means
that it will work depending on what you are querying which is a recipe
for disaster.