Refactor pieces and games helpers to improve readability and fix links
This builds on the awesome work of @radiofreemattd. Switched the pieces
helper to draw the link outside the square so it is clickable when the
square is empty. Moved some of the shared logic into methods that are
shared between the helpers. The selected css is currently broken but this
implementation should be functional.