Enforce turn logic in the controller and fix bug with castling
Since castling involves 2 moves and the turn logic toggled the current
player on every move it broke. Now we just sent the current player to
the opposite color of the last piece moved. Also add a before_action to
the pieces controller with keeps players from moving when not their
turn.